Fix status handling in getRestaurants
The changes in #15 were meant to fix a bug where we were looking at
an incorrect field for the status code of the fetch response.
However, these changes were not thoroughly tested and actually didn't
completely work as coded, because when the response is a success
it does not have a "statusCode" property on the body. This patch
checks the actual HTTP response status instead of the response body
message.