Refactor to add initial React setup
This patch introduces a baseline for using React on the client-side.
There is still work to be done to implement a Flux pattern for data
flow, but the guts are here and working. I intentionally did not setup
hot-loading because that requires socket.io which has a native addon
dependency, which can be a headache to get working on windows. The
primary focus of this patch is:

 - babel ES6 transformation of client-side assets
 - whatwg-fetch polyfill automatically available to client-side js
 - React compilation on `npm start`

Some additional changes made:

 - moved client-side js to lib/client to make the compilation step needed
   explicit.
 - moved server code to lib/server
 - GET /restaurants now returns an id along with the name, to make React
   happy.