Added a bit of configuration checking in the default config.  Adding it into default allows us to easily override the checks for production or development builds if desired (though I'm not sure we would ever actually *want* that..)
Right now the only check is for whether or not the Google Places API key appears to be supplied. We don't even check if the key is valid just that the parameter was supplied.

Removed the now unneccessary check of whether the key was supplied from the restaurants end point.