Simplify the README
This patch is an effort to greatly streamline the readme to the
bare essentials while still providing enough information to give
context and setup guidance. Most of the extra information was moved
to the wiki, but the Docker instructions will need to be resurrected
from the dead and moved to the wiki as well.